CODES/Trumpf_Metamation/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.2.0
fastapi==0.87.0
fastjsonschema==2.16.2
filelock==3.9.0
filetype==1.2.0
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.12.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.26.1
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.10
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.6.0.66
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
pooch==1.6.0
pretrainedmodels==0.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyMatting==1.1.8
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.6
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
rembg==2.0.30
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.21.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
tinycss2==1.2.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
yarl==1.8.2
zipp==3.11.0
